Refactor auto-clicker into modular structure with improved maintainability.
