+++ autoclick/__init__.py
@@ -1,0 +1,2 @@
+# Auto Click Package
+__version__ = '1.0.0'
